1. 시세 캔들 조회(일) api 추가 2. 볼린저 밴드, MACD, 이동 평균선 지수 구하는 공식 추가

+++ app/src/main/java/com/yun/mysimplecoin/util/Util.kt
@@ -67,7 +67,7 @@ object Util {
         return "70"
     }
 
-    fun getStandardDeviation(numbers: List<Int>): Double {
+    fun getStandardDeviation(numbers: List<Double>): Double {
         val size = numbers.size
         val mean = numbers.average()
         val variance = numbers.map { (it - mean) * (it - mean) }.sum() / (size - 1)
